+++ src/tools/check/src/main.rs
@@ -5,7 +5,9 @@ use kime_engine_cffi::{
     Config, InputCategory, InputEngine, InputResult_CONSUMED, InputResult_HAS_COMMIT,
     InputResult_HAS_PREEDIT,
 };
+use kime_version::build::VERSION;
 use pad::PadStr;
+use semver::{Version, VersionReq};
 use std::env;
 use std::io::BufRead;
 use strum::{EnumIter, EnumMessage, IntoEnumIterator, IntoStaticStr};
@@ -135,6 +137,17 @@ impl Check {
                     Ok(config) => config,
                     Err(err) => return CondResult::Fail(format!("Can't parse config.yaml: {err}")),
                 };
+
+                let current_version = Version::parse(VERSION.trim()).unwrap();
+
+                // too low version, for example.
+                if config.kime_version.matches(&Version::new(1, 0, 0)) {
+                    return CondResult::Fail(format!("Version information should be specified and not weak. Put this in the config file: `kime_version: '^{current_version}'`"));
+                }
+
+                if !config.kime_version.matches(&current_version) {
+                    return CondResult::Fail(format!("Current Kime version is {current_version}, does not matches with {}. Please check our CHANGELOG.md, and change to: `kime_version: '^{current_version}'`", config.kime_version));
+                }
 
                 match config.engine.translation_layer {
                     Some(ref raw_path) => {